Moved all game objects into their own package and started work on an AIPlayer
Basically, this commit warped a tiny bit. Originally I was moving tings around and also made the aiblitz.go and blitz.go in the game package use similar code.

Then I started moving files around and started work on an AIPlayer that used the same WriteMessage etc of a normal player. I got pretty clsoe. but didn't figure it out but since so much has changed I wanted to store it on this branch for now